chore: Moves Tooltip to the components folder (#14137)
