+++ superset-frontend/src/SqlLab/components/ColumnElement.tsx
@@ -21,7 +21,7 @@ import PropTypes from 'prop-types';
 import { ClassNames } from '@emotion/core';
 import { styled, useTheme } from '@superset-ui/core';
 
-import { Tooltip } from 'src/common/components/Tooltip';
+import { Tooltip } from 'src/components/Tooltip';
 
 const propTypes = {
   column: PropTypes.object.isRequired,
