+++ superset-frontend/src/components/AlteredSliceTag/index.jsx
@@ -22,7 +22,7 @@ import { isEqual, isEmpty } from 'lodash';
 import { t } from '@superset-ui/core';
 import getControlsForVizType from 'src/utils/getControlsForVizType';
 import { safeStringify } from 'src/utils/safeStringify';
-import { Tooltip } from 'src/common/components/Tooltip';
+import { Tooltip } from 'src/components/Tooltip';
 import ModalTrigger from '../ModalTrigger';
 import TableView from '../TableView';
