+++ superset-frontend/src/components/CachedLabel/index.tsx
@@ -19,7 +19,7 @@
 import React, { useState } from 'react';
 import { t } from '@superset-ui/core';
 import Label from 'src/components/Label';
-import { Tooltip } from 'src/common/components/Tooltip';
+import { Tooltip } from 'src/components/Tooltip';
 import { TooltipContent } from './TooltipContent';
 
 interface Props {
