+++ superset-frontend/src/components/CopyToClipboard/index.jsx
@@ -19,7 +19,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { t } from '@superset-ui/core';
-import { Tooltip } from 'src/common/components/Tooltip';
+import { Tooltip } from 'src/components/Tooltip';
 import withToasts from 'src/messageToasts/enhancers/withToasts';
 import copyTextToClipboard from 'src/utils/copy';
