+++ superset-frontend/src/components/EditableTitle/index.tsx
@@ -19,7 +19,7 @@
 import React, { useEffect, useState, useRef } from 'react';
 import cx from 'classnames';
 import { t } from '@superset-ui/core';
-import { Tooltip } from 'src/common/components/Tooltip';
+import { Tooltip } from 'src/components/Tooltip';
 
 export interface EditableTitleProps {
   canEdit?: boolean;
