+++ superset-frontend/src/components/FaveStar/index.tsx
@@ -19,7 +19,7 @@
 
 import React, { useCallback } from 'react';
 import { t, styled } from '@superset-ui/core';
-import { Tooltip } from 'src/common/components/Tooltip';
+import { Tooltip } from 'src/components/Tooltip';
 import { useComponentDidMount } from 'src/common/hooks/useComponentDidMount';
 import Icon from '../Icon';
