+++ superset-frontend/src/components/IconTooltip/index.tsx
@@ -17,7 +17,7 @@
  * under the License.
  */
 import React, { ReactNode } from 'react';
-import { Tooltip } from 'src/common/components/Tooltip';
+import { Tooltip } from 'src/components/Tooltip';
 import { styled } from '@superset-ui/core';
 
 export interface Props {
