+++ superset-frontend/src/components/ListView/ActionsBar.tsx
@@ -18,7 +18,7 @@
  */
 import React from 'react';
 import { styled } from '@superset-ui/core';
-import { Tooltip } from 'src/common/components/Tooltip';
+import { Tooltip } from 'src/components/Tooltip';
 import { IconName } from 'src/components/Icon';
 import Icons from 'src/components/Icons';
 import { TooltipPlacement } from 'antd/lib/tooltip';
