+++ superset-frontend/src/dashboard/components/PublishedStatus/index.jsx
@@ -19,7 +19,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { t } from '@superset-ui/core';
-import { Tooltip } from 'src/common/components/Tooltip';
+import { Tooltip } from 'src/components/Tooltip';
 import Label from 'src/components/Label';
 
 const propTypes = {
