+++ superset-frontend/src/dashboard/components/nativeFilters/FilterBar/FilterSets/FilterSetUnit.tsx
@@ -22,8 +22,8 @@ import { FilterSet } from 'src/dashboard/reducers/types';
 import { DataMaskState } from 'src/dataMask/types';
 import { CheckOutlined, EllipsisOutlined } from '@ant-design/icons';
 import { HandlerFunction, styled, supersetTheme, t } from '@superset-ui/core';
-import { Tooltip } from 'src/common/components/Tooltip';
 import Button from 'src/components/Button';
+import { Tooltip } from 'src/components/Tooltip';
 import FiltersHeader from './FiltersHeader';
 import { getFilterBarTestId } from '..';
