+++ superset-frontend/src/explore/components/QueryAndSaveBtns.jsx
@@ -21,7 +21,7 @@ import PropTypes from 'prop-types';
 import ButtonGroup from 'src/components/ButtonGroup';
 import { t, useTheme } from '@superset-ui/core';
 
-import { Tooltip } from 'src/common/components/Tooltip';
+import { Tooltip } from 'src/components/Tooltip';
 import Button from 'src/components/Button';
 
 const propTypes = {
