+++ superset-frontend/src/explore/components/controls/VizTypeControl.jsx
@@ -21,8 +21,8 @@ import PropTypes from 'prop-types';
 import { Row, Col, FormControl } from 'react-bootstrap';
 import { Behavior, t, getChartMetadataRegistry } from '@superset-ui/core';
 import { useDynamicPluginContext } from 'src/components/DynamicPlugins';
-import { Tooltip } from 'src/common/components/Tooltip';
 import Modal from 'src/components/Modal';
+import { Tooltip } from 'src/components/Tooltip';
 import Label from 'src/components/Label';
 import ControlHeader from '../ControlHeader';
 import './VizTypeControl.less';
