+++ superset-frontend/src/views/CRUD/data/query/QueryList.tsx
@@ -36,7 +36,7 @@ import ListView, {
   ListViewProps,
 } from 'src/components/ListView';
 import Icon, { IconName } from 'src/components/Icon';
-import { Tooltip } from 'src/common/components/Tooltip';
+import { Tooltip } from 'src/components/Tooltip';
 import SyntaxHighlighter from 'react-syntax-highlighter/dist/cjs/light';
 import sql from 'react-syntax-highlighter/dist/cjs/languages/hljs/sql';
 import github from 'react-syntax-highlighter/dist/cjs/styles/hljs/github';
